pyproject.toml
[tool.poetry]
name = "learn-fastapi"
version = "0.1.0"
description = "Learning FastAPI Project"
authors = ["KKamji <xowl5460@naver.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"} # JWT 및 JWE 토큰 처리
bcrypt = "^4.2.0" # 암호화 라이브러리
python-multipart = "^0.0.17" # 파일 업로드 지원 라이브러리 (Form)

[tool.poetry.group.dev.dependencies]
requests = "2.31.0"
pytest = "^8.3.3"
httpie = "^3.2.4"
httpx = "^0.27.2"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
